requirements.txt
# Airflow
apache-airflow==2.10.5 
--constraint "https://raw.githubusercontent.com/apache/airflow/constraints-2.10.5/constraints-3.10.txt"
apache-airflow-providers-fab
apache-airflow[statsd]

# Others
ibapi==9.81.1.post1
transformers==4.48.0
yfinance==0.2.51
sqlalchemy==1.4.54
pyautogui==0.9.54
opencv-python==4.11.0.86
tensorflow==2.18.0
keras==3.8.0
matplotlib==3.10.0
keras-tuner==1.4.7
scikit-learn==1.5.2
aiomysql
